+++ application/main-application/src/main/java/org/mainapplication/user/service/UserServiceImp.java
@@ -1,11 +1,15 @@
 package org.mainapplication.user.service;
 
+import org.domainmodule.user.entity.Oauth;
 import org.domainmodule.user.entity.User;
+import org.domainmodule.user.entity.type.ProviderType;
+import org.domainmodule.user.repository.OauthRepository;
 import org.domainmodule.user.repository.UserRepository;
 import org.mainapplication.auth.service.AuthService;
 import org.mainapplication.global.oauth2.userinfo.OAuth2UserInfo;
 import org.springframework.stereotype.Service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -14,14 +18,19 @@ public class UserServiceImp implements UserService {
 
 	private final UserRepository userRepository;
 	private final AuthService authService;
+	private final OauthRepository oauthRepository;
 
 	@Override
+	@Transactional
 	public User loginOrCreateUser(OAuth2UserInfo oAuth2Response) {
-		return userRepository.findByEmail(oAuth2Response.getEmail())
+		ProviderType providerType = ProviderType.fromValue(oAuth2Response.getProvider());
+		return oauthRepository.findByProviderAndProviderId(providerType, oAuth2Response.getProviderId())
+			.map(Oauth::getUser)
 			.orElseGet(() -> registerUser(oAuth2Response));
 	}
 
 	@Override
+	@Transactional
 	public User registerUser(OAuth2UserInfo oAuth2Response) {
 		User user = createAndSaveUser(oAuth2Response);
 		authService.createAndSaveOauth(oAuth2Response, user);
