+++ domain/domain-module/src/main/java/org/domainmodule/user/repository/OauthRepository.java
@@ -3,8 +3,9 @@ package org.domainmodule.user.repository;
 import java.util.Optional;
 
 import org.domainmodule.user.entity.Oauth;
+import org.domainmodule.user.entity.type.ProviderType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OauthRepository extends JpaRepository<Oauth, Long> {
-	Optional<Oauth> findByProviderAndProviderId(String provider, String providerId);
+	Optional<Oauth> findByProviderAndProviderId(ProviderType provider, String providerId);
 }
